Refactor GenericMedia.py -> generic.py
