+++ medusa/media/banner.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with Medusa. If not, see <http://www.gnu.org/licenses/>.
 
-from .GenericMedia import GenericMedia
+from .generic import GenericMedia
 from ..image_cache import ImageCache
