+++ pytest.ini
@@ -39,9 +39,9 @@ flake8-ignore =
     medusa/indexers/tvmaze/tvmaze_api.py D100 D101 D102 D103 D105 D200 D202 D204 D205 D210 D400 D401 E711 N802 N803 N806
     medusa/indexers/tvmaze/tvmaze_exceptions.py D200 D204 D205 D400 N801
     medusa/indexers/tvmaze/tvmaze_ui.py D102 D103 D200 D204 D400
-    medusa/media/GenericMedia.py D100 D101 D200 D202 D205 D400
     medusa/media/banner.py D100 D102 D200 D400
     medusa/media/fan_art.py D100 D102 D200 D400
+    medusa/media/generic.py D100 D101 D200 D202 D205 D400
     medusa/media/network_logo.py D100 D102 D200 D400
     medusa/media/poster.py D100 D102 D200 D400
     medusa/media/__init__.py D104
